+++ AppLovin/com/mopub/mobileads/AppLovinRewardedVideo.java
@@ -27,11 +27,23 @@ import com.mopub.common.logging.MoPubLog;
 import java.util.HashMap;
 import java.util.Map;
 
+import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.CLICKED;
+import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.CUSTOM;
+import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.CUSTOM_WITH_THROWABLE;
+import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.LOAD_ATTEMPTED;
+import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.LOAD_FAILED;
+import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.LOAD_SUCCESS;
+import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.SHOULD_REWARD;
+import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.SHOW_ATTEMPTED;
+import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.SHOW_FAILED;
+import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.SHOW_SUCCESS;
+
 public class AppLovinRewardedVideo extends CustomEventRewardedVideo implements AppLovinAdLoadListener, AppLovinAdDisplayListener, AppLovinAdClickListener, AppLovinAdVideoPlaybackListener, AppLovinAdRewardListener {
 
     private static final String DEFAULT_ZONE = "";
     private static final String DEFAULT_TOKEN_ZONE = "token";
     private static final String ZONE_ID_SERVER_EXTRAS_KEY = "zone_id";
+    private static final String ADAPTER_NAME = AppLovinRewardedVideo.class.getSimpleName();
 
     // A map of Zone -> `AppLovinIncentivizedInterstitial` to be shared by instances of the custom event.
     // This prevents skipping of ads as this adapter will be re-created and preloaded (along with underlying `AppLovinIncentivizedInterstitial`)
@@ -51,9 +63,16 @@ public class AppLovinRewardedVideo extends CustomEventRewardedVideo implements A
     private AppLovinAd tokenAd;
     private String serverExtrasZoneId = DEFAULT_ZONE;
 
+    @NonNull
+    private AppLovinAdapterConfiguration mAppLovinAdapterConfiguration;
+
     //
     // MoPub Custom Event Methods
     //
+
+    public AppLovinRewardedVideo() {
+        mAppLovinAdapterConfiguration = new AppLovinAdapterConfiguration();
+    }
 
     @Override
     protected boolean checkAndInitializeSdk(@NonNull final Activity activity, @NonNull final Map<String, Object> localExtras, @NonNull final Map<String, String> serverExtras) throws Exception {
@@ -62,7 +81,7 @@ public class AppLovinRewardedVideo extends CustomEventRewardedVideo implements A
         boolean canCollectPersonalInfo = MoPub.canCollectPersonalInformation();
         AppLovinPrivacySettings.setHasUserConsent(canCollectPersonalInfo, activity.getApplicationContext());
 
-        MoPubLog.d("Initializing AppLovin rewarded video...");
+        MoPubLog.log(CUSTOM, ADAPTER_NAME, "Initializing AppLovin rewarded video...");
 
         if (!initialized) {
             sdk = retrieveSdk(serverExtras, activity);
@@ -84,15 +103,18 @@ public class AppLovinRewardedVideo extends CustomEventRewardedVideo implements A
         final String adMarkup = serverExtras.get(DataKeys.ADM_KEY);
         final boolean hasAdMarkup = !TextUtils.isEmpty(adMarkup);
 
-        MoPubLog.d("Requesting AppLovin banner with serverExtras: " + serverExtras + ", localExtras: " + localExtras + " and has ad markup: " + hasAdMarkup);
+        MoPubLog.log(CUSTOM, ADAPTER_NAME, "Requesting AppLovin rewarded video with serverExtras: " + serverExtras
+                + ", localExtras: " + localExtras + " and has ad markup: " + hasAdMarkup);
+
+        mAppLovinAdapterConfiguration.setCachedInitializationParameters(activity, serverExtras);
 
         // Determine zone
         final String zoneId;
         if (hasAdMarkup) {
             zoneId = DEFAULT_TOKEN_ZONE;
         } else {
-            if (!TextUtils.isEmpty(serverExtras.get(ZONE_ID_SERVER_EXTRAS_KEY))) {
-                serverExtrasZoneId = serverExtras.get(ZONE_ID_SERVER_EXTRAS_KEY);
+            serverExtrasZoneId = serverExtras.get(ZONE_ID_SERVER_EXTRAS_KEY);
+            if (!TextUtils.isEmpty(serverExtrasZoneId)) {
                 zoneId = serverExtrasZoneId;
             } else {
                 zoneId = DEFAULT_ZONE;
@@ -107,15 +129,19 @@ public class AppLovinRewardedVideo extends CustomEventRewardedVideo implements A
             isTokenEvent = true;
 
             sdk.getAdService().loadNextAdForAdToken(adMarkup, this);
+            MoPubLog.log(getAdNetworkId(), LOAD_ATTEMPTED, ADAPTER_NAME);
         }
         // Zone/regular ad load
         else {
             incentivizedInterstitial.preload(this);
+            MoPubLog.log(getAdNetworkId(), LOAD_ATTEMPTED, ADAPTER_NAME);
         }
     }
 
     @Override
     protected void showVideo() {
+        MoPubLog.log(SHOW_ATTEMPTED, ADAPTER_NAME);
+
         if (hasVideoAvailable()) {
             fullyWatched = false;
             reward = null;
@@ -126,7 +152,12 @@ public class AppLovinRewardedVideo extends CustomEventRewardedVideo implements A
                 incentivizedInterstitial.show(parentActivity, null, this, this, this, this);
             }
         } else {
-            MoPubLog.d("Failed to show an AppLovin rewarded video before one was loaded");
+            MoPubLog.log(SHOW_FAILED,
+                    ADAPTER_NAME,
+                    MoPubErrorCode.NETWORK_NO_FILL.getIntCode(),
+                    MoPubErrorCode.NETWORK_NO_FILL);
+
+            MoPubLog.log(CUSTOM, ADAPTER_NAME, "Failed to show an AppLovin rewarded video before one was loaded");
             MoPubRewardedVideoManager.onRewardedVideoPlaybackError(getClass(), getAdNetworkId(), MoPubErrorCode.VIDEO_PLAYBACK_ERROR);
         }
     }
@@ -162,7 +193,7 @@ public class AppLovinRewardedVideo extends CustomEventRewardedVideo implements A
 
     @Override
     public void adReceived(final AppLovinAd ad) {
-        MoPubLog.d("Rewarded video did load ad: " + ad.getAdIdNumber());
+        MoPubLog.log(CUSTOM, ADAPTER_NAME, "Rewarded video did load ad: " + ad.getAdIdNumber());
 
         if (isTokenEvent) {
             tokenAd = ad;
@@ -173,8 +204,11 @@ public class AppLovinRewardedVideo extends CustomEventRewardedVideo implements A
             public void run() {
                 try {
                     MoPubRewardedVideoManager.onRewardedVideoLoadSuccess(AppLovinRewardedVideo.this.getClass(), getAdNetworkId());
+
+                    MoPubLog.log(LOAD_SUCCESS, ADAPTER_NAME);
                 } catch (Throwable th) {
-                    MoPubLog.e("Unable to notify listener of successful ad load.", th);
+                    MoPubLog.log(CUSTOM_WITH_THROWABLE, "Unable to notify listener of " +
+                            "successful ad load.", th);
                 }
             }
         });
@@ -182,15 +216,20 @@ public class AppLovinRewardedVideo extends CustomEventRewardedVideo implements A
 
     @Override
     public void failedToReceiveAd(final int errorCode) {
-        MoPubLog.d("Rewarded video failed to load with error: " + errorCode);
 
         parentActivity.runOnUiThread(new Runnable() {
             @Override
             public void run() {
                 try {
-                    MoPubRewardedVideoManager.onRewardedVideoLoadFailure(AppLovinRewardedVideo.this.getClass(), getAdNetworkId(), toMoPubErrorCode(errorCode));
+                    MoPubRewardedVideoManager.onRewardedVideoLoadFailure(AppLovinRewardedVideo.this.
+                            getClass(), getAdNetworkId(), toMoPubErrorCode(errorCode));
+
+                    MoPubLog.log(LOAD_FAILED, ADAPTER_NAME,
+                            toMoPubErrorCode(errorCode).getIntCode(),
+                            toMoPubErrorCode(errorCode));
                 } catch (Throwable th) {
-                    MoPubLog.e("Unable to notify listener of failure to receive ad.", th);
+                    MoPubLog.log(CUSTOM_WITH_THROWABLE, "Unable to notify listener of failure" +
+                            " to receive ad.", th);
                 }
             }
         });
@@ -202,16 +241,18 @@ public class AppLovinRewardedVideo extends CustomEventRewardedVideo implements A
 
     @Override
     public void adDisplayed(final AppLovinAd ad) {
-        MoPubLog.d("Rewarded video displayed");
         MoPubRewardedVideoManager.onRewardedVideoStarted(getClass(), getAdNetworkId());
+
+        MoPubLog.log(SHOW_SUCCESS, ADAPTER_NAME);
     }
 
     @Override
     public void adHidden(final AppLovinAd ad) {
-        MoPubLog.d("Rewarded video dismissed");
 
         if (fullyWatched && reward != null) {
-            MoPubLog.d("Rewarded" + reward.getAmount() + " " + reward.getLabel());
+            MoPubLog.log(CUSTOM, ADAPTER_NAME, "Rewarded: " + reward.getAmount() + " " + reward.getLabel());
+            MoPubLog.log(SHOULD_REWARD, ADAPTER_NAME, reward.getAmount(), reward.getLabel());
+
             MoPubRewardedVideoManager.onRewardedVideoCompleted(getClass(), getAdNetworkId(), reward);
         }
 
@@ -224,8 +265,9 @@ public class AppLovinRewardedVideo extends CustomEventRewardedVideo implements A
 
     @Override
     public void adClicked(final AppLovinAd ad) {
-        MoPubLog.d("Rewarded video clicked");
         MoPubRewardedVideoManager.onRewardedVideoClicked(getClass(), getAdNetworkId());
+
+        MoPubLog.log(CLICKED, ADAPTER_NAME);
     }
 
     //
@@ -234,12 +276,12 @@ public class AppLovinRewardedVideo extends CustomEventRewardedVideo implements A
 
     @Override
     public void videoPlaybackBegan(final AppLovinAd ad) {
-        MoPubLog.d("Rewarded video playback began");
+        MoPubLog.log(CUSTOM, ADAPTER_NAME, "Rewarded video playback began");
     }
 
     @Override
     public void videoPlaybackEnded(final AppLovinAd ad, final double percentViewed, final boolean fullyWatched) {
-        MoPubLog.d("Rewarded video playback ended at playback percent: " + percentViewed);
+        MoPubLog.log(CUSTOM, ADAPTER_NAME, "Rewarded video playback ended at playback percent: " + percentViewed);
 
         this.fullyWatched = fullyWatched;
     }
@@ -250,22 +292,25 @@ public class AppLovinRewardedVideo extends CustomEventRewardedVideo implements A
 
     @Override
     public void userOverQuota(final AppLovinAd appLovinAd, final Map map) {
-        MoPubLog.d("Rewarded video validation request for ad did exceed quota with response: " + map);
+        MoPubLog.log(CUSTOM, ADAPTER_NAME, "Rewarded video validation request for ad did exceed quota with " +
+                "response: " + map);
     }
 
     @Override
     public void validationRequestFailed(final AppLovinAd appLovinAd, final int errorCode) {
-        MoPubLog.d("Rewarded video validation request for ad failed with error code: " + errorCode);
+        MoPubLog.log(CUSTOM, ADAPTER_NAME, "Rewarded video validation request for ad failed with error " +
+                "code: " + errorCode);
     }
 
     @Override
     public void userRewardRejected(final AppLovinAd appLovinAd, final Map map) {
-        MoPubLog.d("Rewarded video validation request was rejected with response: " + map);
+        MoPubLog.log(CUSTOM, ADAPTER_NAME, "Rewarded video validation request was rejected with response: "
+                + map);
     }
 
     @Override
     public void userDeclinedToViewAd(final AppLovinAd appLovinAd) {
-        MoPubLog.d("User declined to view rewarded video");
+        MoPubLog.log(CUSTOM, ADAPTER_NAME, "User declined to view rewarded video");
         MoPubRewardedVideoManager.onRewardedVideoClosed(getClass(), getAdNetworkId());
     }
 
@@ -274,7 +319,7 @@ public class AppLovinRewardedVideo extends CustomEventRewardedVideo implements A
         final String currency = (String) map.get("currency");
         final int amount = (int) Double.parseDouble((String) map.get("amount")); // AppLovin returns amount as double
 
-        MoPubLog.d("Verified " + amount + " " + currency);
+        MoPubLog.log(CUSTOM, ADAPTER_NAME, "Verified " + amount + " " + currency);
 
         reward = MoPubReward.success(currency, amount);
     }