+++ AppLovin/CHANGELOG.md
@@ -1,4 +1,12 @@
 ## Changelog
+  * 9.1.3.1
+    * **Note**: This version is only compatible with the 5.5.0+ release of the MoPub SDK.
+    * Add the `AppLovinAdapterConfiguration` class to: 
+         * pre-initialize the AppLovin SDK during MoPub SDK initialization process
+         * store adapter and SDK versions for logging purpose
+         * return the Advanced Biding token previously returned by `AppLovinAdvancedBidder.java`
+    * Streamline adapter logs via `MoPubLog` to make debugging more efficient. For more details, check the [Android Initialization guide](https://developers.mopub.com/docs/android/initialization/) and [Writing Custom Events guide](https://developers.mopub.com/docs/android/custom-events/).
+
   * 9.1.3.0
     * This version of the adapters has been certified with AppLovin 9.1.3.